Use archive tar.gz instead of RPM in Presto container
Move the artifact (archive & CLI) download outside of the Presto Dockerfile.
Introduce a `jenkins-release-image.sh` script that can drive the build
and push of new containers. This will replace the inline script
currently used in the Jenkins job.

A pleasant side effect of this change is that the resulting
container's size goes 3.43 GB to 1.17 GB.